+++ packages/ttsc/test/cli/api_compile_and_transform_test.go
@@ -11,10 +11,10 @@ import (
 // The API commands expose compiler results as JSON for the JavaScript wrapper.
 // They must preserve project-relative keys while avoiding writes to disk.
 //
-// 1. Materialize a compilable project with one TypeScript source file.
-// 2. Run `api-compile` through the real `cmd/ttsc` front door.
-// 3. Run `api-transform` and assert the source-text DTO uses project-relative
-//    keys just like the JavaScript wrapper expects.
+//  1. Materialize a compilable project with one TypeScript source file.
+//  2. Run `api-compile` through the real `cmd/ttsc` front door.
+//  3. Run `api-transform` and assert the source-text DTO uses project-relative
+//     keys just like the JavaScript wrapper expects.
 func TestCLIAPICompileAndTransform(t *testing.T) {
 	root := t.TempDir()
