+++ packages/ttsc/test/driver/splice_call_ignores_closing_paren_inside_line_comment_test.go
@@ -5,13 +5,9 @@ import "testing"
 // TestDriverSpliceCallIgnoresClosingParenInsideLineComment verifies line
 // comments do not terminate call scanning.
 //
-// The scanner must skip line-comment text until the next newline before it
-// resumes parenthesis accounting. Otherwise a commented `)` can truncate the
-// rewrite range and leave malformed TypeScript behind.
-//
-// This scenario targets the line-comment branch of the private splice helper.
-// The input spans multiple lines to prove scanning resumes after the comment
-// before closing the original call.
+// This scenario targets the line-comment branch of the private splice helper
+// because a commented `)` can truncate the rewrite range unless scanning resumes
+// after the newline before closing the original call.
 //
 // 1. Splice a plugin call with a line comment inside the argument list.
 // 2. Include a closing parenthesis character in that line comment.
