+++ packages/ttsc/test/driver/splice_call_keeps_division_outside_regex_mode_test.go
@@ -5,13 +5,9 @@ import "testing"
 // TestDriverSpliceCallKeepsDivisionOutsideRegexMode verifies division
 // operators are not parsed as regex literals.
 //
-// The scanner distinguishes slash tokens that start regular expressions from
-// slash tokens that divide expressions. Division must stay in normal expression
-// mode so the following comma and closing parenthesis are handled correctly.
-//
-// This scenario exercises the non-regex slash path through the splice helper.
-// It preserves coverage for a subtle lexer boundary without relocating the
-// test back into the production driver package.
+// This scenario exercises the non-regex slash path through the splice helper
+// because division must stay in normal expression mode so the following comma
+// and closing parenthesis are handled correctly.
 //
 // 1. Splice a plugin call whose first argument uses division.
 // 2. Consume the call parentheses through the driver rewrite helper.
